fix: support undo/redo/selectAll on WebIDE textInput
