+++ packages/core-browser/src/common/common.contribution.ts
@@ -20,7 +20,7 @@ import { MenuId } from '../menu/next/menu-id';
 import { PreferenceContribution } from '../preferences';
 import { AppConfig } from '../react-providers/config-provider';
 
-import { FILE_COMMANDS, COMMON_COMMANDS, EDITOR_COMMANDS, TERMINAL_COMMANDS, LAYOUT_COMMANDS } from './common.command';
+import { FILE_COMMANDS, COMMON_COMMANDS, EDITOR_COMMANDS, TERMINAL_COMMANDS } from './common.command';
 import { ClientAppContribution } from './common.define';
 
 export const inputFocusedContextKey = 'inputFocus';
