+++ packages/monaco/src/common/command.ts
@@ -1,0 +1,5 @@
+export const DELEGATE_COMMANDS = {
+  UNDO: 'undo',
+  REDO: 'redo',
+  SELECT_ALL: 'editor.action.selectAll',
+};
